+++ core/build.gradle
@@ -39,6 +39,7 @@ dependencies {
   implementation 'org.hyperledger.besu:plugin-api'
   implementation 'org.jdbi:jdbi3-core'
   implementation 'tech.pegasys.signers.internal:bls-keystore'
+  implementation 'tech.pegasys.signers.internal:keystorage-aws'
   implementation 'tech.pegasys.signers.internal:keystorage-azure'
   implementation 'tech.pegasys.signers.internal:keystorage-hashicorp'
   implementation 'tech.pegasys.signers.internal:signing-secp256k1-api'
