+++ gradle/versions.gradle
@@ -97,7 +97,7 @@ dependencyManagement {
 
     dependency 'tech.pegasys:jblst:0.3.6-4'
 
-    dependencySet(group: 'tech.pegasys.signers.internal', version: '2.1.0') {
+    dependencySet(group: 'tech.pegasys.signers.internal', version: 'develop') {
       entry 'bls-keystore'
       entry 'keystorage-hashicorp'
       entry 'keystorage-azure'
