+++ signing/src/main/java/tech/pegasys/web3signer/signing/config/metadata/AbstractArtifactSignerFactory.java
@@ -16,6 +16,7 @@ import static java.lang.String.format;
 import static java.nio.charset.StandardCharsets.UTF_8;
 
 import tech.pegasys.signers.aws.AwsSecretsManager;
+import tech.pegasys.signers.aws.AwsSecretsManagerProvider;
 import tech.pegasys.signers.azure.AzureKeyVault;
 import tech.pegasys.signers.hashicorp.HashicorpConnection;
 import tech.pegasys.signers.hashicorp.HashicorpConnectionFactory;
@@ -98,16 +99,6 @@ public abstract class AbstractArtifactSignerFactory implements ArtifactSignerFac
       throw new SigningMetadataException(
           "Failed to fetch secret from Interlock: " + e.getMessage(), e);
     }
-  }
-
-  protected Bytes extractBytesFromSecretsManager(final AwsKeySigningMetadata metadata) {
-    final AwsSecretsManager awsSecretsManager =
-        AwsSecretsManagerFactory.createAwsSecretsManager(metadata);
-    return awsSecretsManager
-        .fetchSecret(metadata.getSecretName())
-        .map(Bytes::fromHexString)
-        .orElseThrow(
-            () -> new SigningMetadataException("Failed to fetch secret from AWS Secrets Manager"));
   }
 
   protected Bytes extractOpaqueDataFromYubiHsm(
