+++ signing/src/main/java/tech/pegasys/web3signer/signing/config/metadata/ArtifactSignerFactory.java
@@ -12,6 +12,7 @@
  */
 package tech.pegasys.web3signer.signing.config.metadata;
 
+import tech.pegasys.signers.aws.AwsSecretsManagerProvider;
 import tech.pegasys.web3signer.signing.ArtifactSigner;
 
 public interface ArtifactSignerFactory {
