+++ signing/src/main/java/tech/pegasys/web3signer/signing/config/metadata/BlsArtifactSignerFactory.java
@@ -12,6 +12,8 @@
  */
 package tech.pegasys.web3signer.signing.config.metadata;
 
+import tech.pegasys.signers.aws.AwsSecretsManager;
+import tech.pegasys.signers.aws.AwsSecretsManagerProvider;
 import tech.pegasys.signers.bls.keystore.KeyStore;
 import tech.pegasys.signers.bls.keystore.KeyStoreLoader;
 import tech.pegasys.signers.bls.keystore.KeyStoreValidationException;
@@ -22,6 +24,7 @@ import tech.pegasys.teku.bls.BLSSecretKey;
 import tech.pegasys.web3signer.common.Web3SignerMetricCategory;
 import tech.pegasys.web3signer.signing.ArtifactSigner;
 import tech.pegasys.web3signer.signing.KeyType;
+import tech.pegasys.web3signer.signing.config.AwsSecretsManagerFactory;
 import tech.pegasys.web3signer.signing.config.metadata.interlock.InterlockKeyProvider;
 import tech.pegasys.web3signer.signing.config.metadata.yubihsm.YubiHsmOpaqueDataProvider;
 
@@ -40,6 +43,7 @@ public class BlsArtifactSignerFactory extends AbstractArtifactSignerFactory {
 
   private final LabelledMetric<OperationTimer> privateKeyRetrievalTimer;
   private final Function<BlsArtifactSignerArgs, ArtifactSigner> signerFactory;
+  private final AwsSecretsManagerProvider awsSecretsManagerProvider;
 
   public BlsArtifactSignerFactory(
       final Path configsDirectory,
@@ -47,6 +51,7 @@ public class BlsArtifactSignerFactory extends AbstractArtifactSignerFactory {
       final HashicorpConnectionFactory connectionFactory,
       final InterlockKeyProvider interlockKeyProvider,
       final YubiHsmOpaqueDataProvider yubiHsmOpaqueDataProvider,
+      final AwsSecretsManagerProvider awsSecretsManagerProvider,
       final Function<BlsArtifactSignerArgs, ArtifactSigner> signerFactory) {
     super(connectionFactory, configsDirectory, interlockKeyProvider, yubiHsmOpaqueDataProvider);
     privateKeyRetrievalTimer =
@@ -56,6 +61,7 @@ public class BlsArtifactSignerFactory extends AbstractArtifactSignerFactory {
             "Time taken to retrieve private key",
             "signer");
     this.signerFactory = signerFactory;
+    this.awsSecretsManagerProvider = awsSecretsManagerProvider;
   }
 
   @Override
@@ -138,6 +144,16 @@ public class BlsArtifactSignerFactory extends AbstractArtifactSignerFactory {
     } catch (final KeyStoreValidationException e) {
       throw new SigningMetadataException(e.getMessage(), e);
     }
+  }
+
+  protected Bytes extractBytesFromSecretsManager(final AwsKeySigningMetadata metadata) {
+    final AwsSecretsManager awsSecretsManager =
+      AwsSecretsManagerFactory.createAwsSecretsManager(awsSecretsManagerProvider, metadata);
+    return awsSecretsManager
+      .fetchSecret(metadata.getSecretName())
+      .map(Bytes::fromHexString)
+      .orElseThrow(
+        () -> new SigningMetadataException("Failed to fetch secret from AWS Secrets Manager"));
   }
 
   @Override
